Move ODP helper functions out to their own package
